Create category layers
Create an object mkLayers, with the coords and the id of the geoObject. Add method showMarkersLayer in leaflet directive to show markers of the specific category.Still not working: the marker info gets duplicated when browsing several times to the same geoObject. When clicking on marker, details displayed are of the last geoObject retrieved not the one clicked.